server/app/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-environ==0.11.2
psycopg2==2.9.9
uuid6==2024.7.10
pillow==11.0.0
django-cors-headers

# dev
ruff==0.7.0
